projects/rss-reader/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
feedparser==6.0.10
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
APScheduler==3.10.4
gunicorn==21.2.0
openai==0.28.1 